Created upstart services, small improvements
- only wget maven if not installed already
- more exact results from the simulation

+++ haproxy-vm/provision.sh
@@ -8,7 +8,7 @@ echo ">>> Installing packages..."
 apt-get -yq install haproxy openjdk-7-jre-headless openjdk-7-jdk curl > /dev/null
 
 echo ">>> Config, enable, and start haproxy"
-cp -r /vagrant/files/etc/haproxy /etc/haproxy
+cp -r /vagrant/files/etc/haproxy /etc
 cp /vagrant/files/etc/default/haproxy /etc/default/haproxy # necessary to enable it
 service haproxy restart
 
@@ -17,5 +17,10 @@ mkdir -p /opt/myapp
 chown -R vagrant /opt/myapp
 
 echo ">>> Preparing maven..."
-wget http://apache.vianett.no/maven/maven-3/3.1.0/binaries/apache-maven-3.1.0-bin.tar.gz -O /tmp/maven.tgz
-tar xzvf /tmp/maven.tgz -C /vagrant
+if [ ! -f "/vagrant/apache-maven-3.1.0/bin/mvn" ]; then
+  wget http://apache.vianett.no/maven/maven-3/3.1.0/binaries/apache-maven-3.1.0-bin.tar.gz -O /tmp/maven.tgz
+  tar xzvf /tmp/maven.tgz -C /vagrant
+fi
+
+# Setup the services
+cp -vr /vagrant/files/etc/init /etc
